refactor(parser): preprocess file inclusions
start with a first pass of parsing the document to detect
file inclusions and process accordingly (inclusing applying
the level offsets on sections when needed)
