+++ pkg/parser/document_preprocessing_include_files.go
@@ -1,0 +1,435 @@
+package parser
+
+import (
+	"bufio"
+	"bytes"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+
+	"github.com/bytesparadise/libasciidoc/pkg/types"
+
+	"github.com/davecgh/go-spew/spew"
+	"github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
+)
+
+func processFileInclusions(ctx *ParseContext, source io.Reader) (string, error) {
+	if log.IsLevelEnabled(log.DebugLevel) {
+		log.Debugf("processing file inclusions in %s with leveloffset=%s", ctx.filename, spew.Sdump(ctx.levelOffsets))
+	}
+	content := &builder{}
+	scanner := bufio.NewScanner(source)
+	t := newBlockDelimiterTracker()
+	for scanner.Scan() {
+		line := scanner.Bytes()
+		element, err := Parse("", line, append(ctx.Opts, Entrypoint("DocumentRawLine"))...)
+		if err != nil {
+			// content of line was not relevant in the context of preparsing (ie, it's a regular line), so let's keep it as-is
+			content.Write(line)
+		} else {
+			if log.IsLevelEnabled(log.DebugLevel) {
+				log.Debugf("checking element of type '%T'", element)
+			}
+			switch e := element.(type) {
+			case *types.AttributeDeclaration:
+				ctx.attributes.set(e.Name, e.Value)
+				t, _ := e.RawText()
+				content.WriteString(t)
+			case *types.AttributeReset:
+				ctx.attributes.unset(e.Name)
+				t, _ := e.RawText()
+				content.WriteString(t)
+			case *types.RawSection:
+				content.WriteString(ctx.levelOffsets.apply(e))
+			case *types.FileInclusion:
+				f, err := includeFile(ctx.Clone(), e)
+				if err != nil {
+					return "", err
+				}
+				content.WriteString(f)
+			case *types.BlockDelimiter:
+				t.push(types.BlockDelimiterKind(e.Kind))
+				ctx.Opts = append(ctx.Opts, GlobalStore(delimitedBlockScopeKey, t.withinDelimitedBlock()))
+				t, _ := e.RawText()
+				content.WriteString(t)
+			default:
+				return "", fmt.Errorf("unexpected type of element while preprocessinh document: '%T'", e)
+			}
+		}
+	}
+	return content.String(), nil
+}
+
+// replace the content of this FileInclusion element the content of the target file
+// note: there is a trade-off here: we include the whole content of the file in the current
+// fragment, making it potentially big, but at the same time we ensure that the context
+// of the inclusion (for example, within a delimited block) is not lost.
+func includeFile(ctx *ParseContext, incl *types.FileInclusion) (string, error) {
+	if err := applySubstitutionsOnBlockWithLocation(ctx, incl); err != nil {
+		return "", errors.Errorf("Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	}
+	content, err := contentOf(ctx, incl)
+	if err != nil {
+		return "", err
+	}
+	ctx.Opts = append(ctx.Opts, sectionEnabled())
+	return processFileInclusions(ctx, bytes.NewReader(content))
+}
+
+type builder struct {
+	strings.Builder
+	insertLF bool
+}
+
+func (b *builder) WriteString(s string) {
+	b.doInsertLF()
+	b.Builder.WriteString(s)
+}
+
+func (b *builder) Write(p []byte) {
+	b.doInsertLF()
+	b.Builder.Write(p)
+}
+
+func (b *builder) doInsertLF() {
+	if b.insertLF {
+		b.Builder.WriteString("\n")
+	}
+	// from now on, we will insert a `\n` before each new line (but there will be no extra `\n` at the end of the content)
+	b.insertLF = true
+}
+
+func contentOf(ctx *ParseContext, incl *types.FileInclusion) ([]byte, error) {
+	path := incl.Location.Stringify()
+	currentDir := filepath.Dir(ctx.filename)
+	filename := filepath.Join(currentDir, path)
+
+	f, absPath, closeFile, err := open(filename)
+	defer closeFile()
+	if err != nil {
+		return nil, errors.Wrapf(err, "Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	}
+	result := bytes.NewBuffer(nil)
+	scanner := bufio.NewScanner(bufio.NewReader(f))
+	if log.IsLevelEnabled(log.DebugLevel) {
+		log.Debugf("reading %s", filename)
+	}
+	if lr, ok, err := lineRanges(incl); err != nil {
+		return nil, errors.Wrapf(err, "Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	} else if ok {
+		if err := readWithinLines(scanner, result, lr); err != nil {
+			return nil, errors.Wrapf(err, "Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+		}
+	} else if tr, ok, err := tagRanges(incl); err != nil {
+		return nil, errors.Wrapf(err, "Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	} else if ok {
+		if err := readWithinTags(path, scanner, result, tr); err != nil {
+			return nil, errors.Wrapf(err, "Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+		}
+	} else {
+		if err := readAll(scanner, result); err != nil {
+			log.Error(err)
+			return nil, errors.Errorf("Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+		}
+	}
+	if err := scanner.Err(); err != nil {
+		log.Error(err)
+		return nil, errors.Errorf("Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	}
+	// cloning the context to avoid altering the original as we process recursively embedded file inclusions
+	ctx.filename = absPath
+	// if the file to include is not an Asciidoc document, just return the content as "raw lines"
+	if !IsAsciidoc(absPath) {
+		log.Debugf("file '%s' is not an Asciidoc file. Scanning content without looking for nested file inclusions", absPath)
+		// send(types.NewDocumentFragment(lineOffset, []interface{}{types.RawLine(content.String())}), done, resultStream)
+		// return
+		ctx.Opts = append(ctx.Opts, DisableRule(FileInclusion))
+	}
+
+	// level offset
+	// parse the content, and returns the corresponding elements
+	if lvl, found, err := incl.Attributes.GetAsString(types.AttrLevelOffset); err != nil {
+		log.Error(err)
+		return nil, errors.Errorf("Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+	} else if found {
+		offset, err := strconv.Atoi(lvl)
+		if err != nil {
+			log.Error(err)
+			return nil, errors.Errorf("Unresolved directive in %s - %s", ctx.filename, incl.RawText)
+		}
+		if strings.HasPrefix(lvl, "+") || strings.HasPrefix(lvl, "-") {
+			ctx.levelOffsets = append(ctx.levelOffsets, relativeOffset(offset))
+		} else {
+			ctx.levelOffsets = []*levelOffset{absoluteOffset(offset)}
+		}
+	}
+	// if log.IsLevelEnabled(log.DebugLevel) {
+	// 	log.Debugf("content of '%s':\n%s", absPath, result.String())
+	// }
+	return result.Bytes(), nil
+}
+
+type levelOffsets []*levelOffset
+
+func (l levelOffsets) apply(s *types.RawSection) string {
+	for _, offset := range l {
+		offset.apply(s)
+	}
+	if log.IsLevelEnabled(log.DebugLevel) {
+		log.Debugf("applied offsets to section: level is now %d", s.Level)
+	}
+	return s.Stringify()
+}
+
+func (l levelOffsets) clone() levelOffsets {
+	result := make([]*levelOffset, len(l))
+	copy(result, l)
+	return result
+}
+
+// levelOffset a func that applies a given offset to the sections of a child document to include in a parent doc (the caller)
+type levelOffset struct {
+	absolute bool
+	value    int
+}
+
+func (l *levelOffset) apply(s *types.RawSection) {
+	// also, absolute offset becomes relative offset after processing the first section,
+	// so that the hierarchy of subsequent sections of the doc to include is preserved
+	if l.absolute {
+		l.absolute = false
+		l.value = l.value - s.Level
+	}
+	s.OffsetLevel(l.value)
+}
+
+func relativeOffset(offset int) *levelOffset {
+	return &levelOffset{
+		absolute: false,
+		value:    offset,
+	}
+}
+
+func absoluteOffset(offset int) *levelOffset {
+	return &levelOffset{
+		absolute: true,
+		value:    offset,
+	}
+}
+
+func (l levelOffset) String() string {
+	if l.absolute {
+		return strconv.Itoa(l.value)
+	}
+	return fmt.Sprintf("+%d", l.value)
+}
+
+// lineRanges parses the `lines` attribute if it exists in the given FileInclusion, and returns
+// a corresponding `LineRanges` (or `false` if parsing failed to invalid input)
+func lineRanges(incl *types.FileInclusion) (types.LineRanges, bool, error) {
+	lineRanges, exists, err := incl.Attributes.GetAsString(types.AttrLineRanges)
+	if err != nil {
+		return types.LineRanges{}, false, err
+	}
+	if exists {
+		lr, err := Parse("", []byte(lineRanges), Entrypoint("LineRanges"))
+		if err != nil {
+			return types.LineRanges{}, false, err
+		}
+		if log.IsLevelEnabled(log.DebugLevel) {
+			log.Debugf("line ranges to include: %s", spew.Sdump(lr))
+		}
+		return types.NewLineRanges(lr), true, nil
+	}
+	return types.LineRanges{}, false, nil
+}
+
+// tagRanges parses the `tags` attribute if it exists in the given FileInclusion, and returns
+// a corresponding `TagRanges` (or `false` if parsing failed to invalid input)
+func tagRanges(incl *types.FileInclusion) (types.TagRanges, bool, error) {
+	tagRanges, exists, err := incl.Attributes.GetAsString(types.AttrTagRanges)
+	if err != nil {
+		return types.TagRanges{}, false, err
+	}
+	if exists {
+		log.Debugf("tag ranges to include: %v", spew.Sdump(tagRanges))
+		tr, err := Parse("", []byte(tagRanges), Entrypoint("TagRanges"))
+		if err != nil {
+			return types.TagRanges{}, false, err
+		}
+		return types.NewTagRanges(tr), true, nil
+	}
+	return types.TagRanges{}, false, nil
+}
+
+// TODO: instead of reading and parsing afterwards, simply parse the lines immediately? ie: `readWithinLines` -> `parseWithinLines`
+// (also, use a specific entrypoint if the doc is not a .adoc)
+func readWithinLines(scanner *bufio.Scanner, content *bytes.Buffer, lineRanges types.LineRanges) error {
+	line := 0
+	for scanner.Scan() {
+		line++
+		// parse the line in search for the `tag::<tag>[]` or `end:<tag>[]` macros
+		l, err := Parse("", scanner.Bytes(), Entrypoint("IncludedFileLine"))
+		if err != nil {
+			return err
+		}
+		fl, ok := l.(types.IncludedFileLine)
+		if !ok {
+			return errors.Errorf("unexpected type of parsed line in file to include: %T", l)
+		}
+		// skip if the line has tags
+		if fl.HasTag() {
+			continue
+		}
+		// TODO: stop reading if current line above highest range
+		if lineRanges.Match(line) {
+			_, err := content.Write(scanner.Bytes())
+			if err != nil {
+				return err
+			}
+			_, err = content.WriteString("\n")
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func readWithinTags(path string, scanner *bufio.Scanner, content *bytes.Buffer, expectedRanges types.TagRanges) error {
+	// log.Debugf("limiting to tag ranges: %v", expectedRanges)
+	currentRanges := make(map[string]*types.CurrentTagRange, len(expectedRanges)) // ensure capacity
+	lineNumber := 0
+	for scanner.Scan() {
+		lineNumber++
+		line := scanner.Bytes()
+		// parse the line in search for the `tag::<tag>[]` or `end:<tag>[]` macros
+		l, err := Parse("", line, Entrypoint("IncludedFileLine"))
+		if err != nil {
+			return err
+		}
+		fl, ok := l.(types.IncludedFileLine)
+		if !ok {
+			return errors.Errorf("unexpected type of parsed line in file to include: %T", l)
+		}
+		// check if a start or end tag was found in the line
+		if startTag, ok := fl.GetStartTag(); ok {
+			currentRanges[startTag.Value] = &types.CurrentTagRange{
+				StartLine: lineNumber,
+				EndLine:   -1,
+			}
+		}
+		if endTag, ok := fl.GetEndTag(); ok {
+			currentRanges[endTag.Value].EndLine = lineNumber
+		}
+		if expectedRanges.Match(lineNumber, currentRanges) && !fl.HasTag() {
+			_, err := content.Write(scanner.Bytes())
+			if err != nil {
+				return err
+			}
+			_, err = content.WriteString("\n")
+			if err != nil {
+				return err
+			}
+		}
+	}
+	// after the file has been processed, let's check if all tags were "found"
+	for _, tag := range expectedRanges {
+		// log.Debugf("checking if tag '%s' was found...", tag.Name)
+		switch tag.Name {
+		case "*", "**":
+			continue
+		default:
+			if tr, found := currentRanges[tag.Name]; !found {
+				return fmt.Errorf("tag '%s' not found in file to include", tag.Name)
+			} else if tr.EndLine == -1 {
+				log.Warnf("detected unclosed tag '%s' starting at line %d of include file: %s", tag.Name, tr.StartLine, path)
+			}
+		}
+	}
+	return nil
+}
+
+func readAll(scanner *bufio.Scanner, content *bytes.Buffer) error {
+	for scanner.Scan() {
+		// parse the line in search for the `tag::<tag>[]` or `end:<tag>[]` macros
+		l, err := Parse("", scanner.Bytes(), Entrypoint("IncludedFileLine"))
+		if err != nil {
+			return err
+		}
+		fl, ok := l.(types.IncludedFileLine)
+		if !ok {
+			return errors.Errorf("unexpected type of parsed line in file to include: %T", l)
+		}
+		// skip if the line has tags
+		if fl.HasTag() {
+			continue
+		}
+		_, err = content.Write(scanner.Bytes())
+		if err != nil {
+			return err
+		}
+		_, err = content.WriteString("\n")
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func open(path string) (*os.File, string, func(), error) {
+	wd, err := os.Getwd()
+	if err != nil {
+		return nil, "", func() {}, err
+	}
+	absPath, err := filepath.Abs(path)
+	if err != nil {
+		return nil, "", func() {
+			log.Debugf("restoring current working dir to: %s", wd)
+			if err := os.Chdir(wd); err != nil { // restore the previous working directory
+				log.WithError(err).Error("failed to restore previous working directory")
+			}
+		}, err
+	}
+	dir := filepath.Dir(absPath)
+	// TODO: we could skip the Chdir part if we retain the absPath in the context,
+	// and use `filepath.Join` to compute the abspath of the file to include
+	if err = os.Chdir(dir); err != nil {
+		return nil, "", func() {
+			log.Debugf("restoring current working dir to: %s", wd)
+			if err := os.Chdir(wd); err != nil { // restore the previous working directory
+				log.WithError(err).Error("failed to restore previous working directory")
+			}
+		}, err
+	}
+	// read the file per-se
+	// log.Debugf("opening '%s'", absPath)
+	f, err := os.Open(absPath)
+	if err != nil {
+		return nil, absPath, func() {
+			log.Debugf("restoring current working dir to: %s", wd)
+			if err := os.Chdir(wd); err != nil { // restore the previous working directory
+				log.WithError(err).Error("failed to restore previous working directory")
+			}
+		}, err
+	}
+	return f, absPath, func() {
+		log.Debugf("restoring current working dir to: %s", wd)
+		if err := os.Chdir(wd); err != nil { // restore the previous working directory
+			log.WithError(err).Error("failed to restore previous working directory")
+		}
+		if err := f.Close(); err != nil {
+			log.WithError(err).Errorf("failed to close file '%s'", absPath)
+		}
+	}, nil
+}
+
+// IsAsciidoc returns true if the file to include is an asciidoc file (based on the file location extension)
+func IsAsciidoc(path string) bool {
+	ext := filepath.Ext(path)
+	return ext == ".asciidoc" || ext == ".adoc" || ext == ".ad" || ext == ".asc" || ext == ".txt"
+}
